requirements.txt
Appium-Python-Client==3.1.0
pytest==8.3.5
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
Pillow==10.3.0
